requirements.txt
python-telegram-bot
openai
faiss-cpu
sentence-transformers
numpy